% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check2Samples.R
\name{check2Samples}
\alias{check2Samples}
\title{Check overlaps between two samples}
\usage{
check2Samples(meta, id1, id2, data_dir)
}
\arguments{
\item{meta}{Meta table}

\item{id1}{The first subject}

\item{id2}{The second subject}
}
\value{
A matrix that contains
}
\description{
Given two subject's names, the function will check the overlaps between all their
samples.
}
\examples{
# check2Samples(meta, id1="P01", id2="P02")

}
