% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCrosscohort.R
\name{checkCrosscohort}
\alias{checkCrosscohort}
\title{Title Generate oncoprint for cross-cohort overlap}
\usage{
checkCrosscohort(meta, dir)
}
\arguments{
\item{meta}{A meta table that contains all samples' names and the corresponding
patients' ids. The first column should be sample name, and second column should
be subjects' id}

\item{dir}{Data path}
}
\value{
A OncoPrint and a table presenting sequence prevalence rate among the subjects.
}
\description{
Title Generate oncoprint for cross-cohort overlap
}
\examples{
# crossc_overlap <- checkSeq(meta, "./")

}
