% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPairwise.R
\name{checkPairwise}
\alias{checkPairwise}
\title{Title Plot heatmap for pairwise TCR overlaps}
\usage{
checkPairwise(meta, dir)
}
\arguments{
\item{meta}{A meta table that contains all samples' names and the corresponding
patients' ids. The first column should be sample name, and second column should
be subjects' id}

\item{dir}{Data path}
}
\value{
A heatmap and a table presenting pairwise TCR overlaps between any 2 subjects.
}
\description{
The function is used to plot heatmap for pairwise TCR overlaps among
the subjects. Each subject can have multiple samples. It may take a while if
data is big.
}
\examples{
# pairwise_overlap <- checkSeq(meta, "./")

}
