% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Identify outliers}
\usage{
detectOutliers(vec, iters, c, jags_inits, jags_params, jags_model)
}
\arguments{
\item{vec}{Named vector of the pairwise}

\item{iters}{Number of iterations}

\item{c}{Quantile threshold}

\item{jags_inits}{Inits of the parameters}

\item{jags_params}{Parameters of the hierarchical model}

\item{jags_model}{The hierarchical model}
}
\description{
The function is used to identify the outliers
}
