% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetOneSumTable}
\alias{GetOneSumTable}
\title{Title Summary all sequences in a table for a patient}
\usage{
GetOneSumTable(thisID, meta, path)
}
\arguments{
\item{thisID}{Patient ID}

\item{meta}{A meta table that contains all patients' id and the corresponding
samples' names. The first column should be sample names, and second column should
be subjects' ids.}
}
\value{
A dataframe that contains sequences, sequence counts and sequence frequencies.
}
\description{
Each patient can have multiple samples. The data should have 3 columns:
sequences, sequence counts and sequences frequencies.
}
