% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcalculator.R
\name{Pcalculator}
\alias{Pcalculator}
\title{Identify crosscohort overlap outliers}
\usage{
Pcalculator(ref, vec)
}
\arguments{
\item{ref}{Reference vector}

\item{vec}{Your own vector}
}
\description{
The function is used to identify crosscohort overlap outliers.
}
