% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadOnePat}
\alias{ReadOnePat}
\title{Title Read data in tsv format for one patient.}
\usage{
ReadOnePat(thismeta, data_dir)
}
\arguments{
\item{thismeta}{A meta table contains the name of the samples for the patient.}

\item{data_dir}{Folder location for data.}
}
\value{
A dataframe that contains sequences, sequence counts and sequence frequencies.
}
\description{
Each patient can have multiple samples. The data should have 3 columns:
sequences, sequence counts and sequences frequencies.
}
