% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssampleIndicator.R
\name{ssampleIndicator}
\alias{ssampleIndicator}
\title{Title Small sample indicator}
\usage{
ssampleIndicator(meta, dir, low.seq = 1000)
}
\arguments{
\item{meta}{A meta table that contains all samples' names and the corresponding
patients' ids. The first column should be sample name, and second column should
be subjects' id}

\item{dir}{Data path}

\item{low.seq}{A integer that define "low frequency" for each subject. 
If a subject has sequences lower than or equal to this number, it will be noted as "Ouch!"; 
Otherwise it will show NA.}
}
\description{
The function is used to identify if a subject has small number of 
frequencies.
}
\examples{
# 

}
